# openapi/nw-trails-v1.yaml
openapi: 3.0.3
info:
  title: NW Trails Backend API
  version: 1.0.0
  description: |
    API contract for CSIS4280 Group 06 NW Trails backend.
    This contract supports Flutter app integration for landmarks,
    guided routes, location-based check-in, awards/progress, and auth.

servers:
  - url: http://localhost:8080/api/v1
    description: Local development

tags:
  - name: Auth
  - name: Landmarks
  - name: Routes
  - name: CheckIns
  - name: Progress
  - name: Admin

paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in and get access/refresh tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /landmarks:
    get:
      tags: [Landmarks]
      summary: Get landmarks list
      operationId: listLandmarks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            $ref: '#/components/schemas/LandmarkCategory'
          required: false
        - in: query
          name: q
          schema:
            type: string
          required: false
          description: Search keyword in name/address
      responses:
        '200':
          description: Landmarks list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Landmark'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /landmarks/{landmarkId}:
    get:
      tags: [Landmarks]
      summary: Get landmark detail
      operationId: getLandmarkById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LandmarkIdPath'
      responses:
        '200':
          description: Landmark detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Landmark'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /routes:
    get:
      tags: [Routes]
      summary: Get route plans
      operationId: listRoutes
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: difficulty
          schema:
            type: string
            enum: [Easy, Medium, Hard]
          required: false
      responses:
        '200':
          description: Route list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoutePlan'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /routes/{routeId}:
    get:
      tags: [Routes]
      summary: Get route detail
      operationId: getRouteById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RouteIdPath'
      responses:
        '200':
          description: Route detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePlan'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /routes/{routeId}/start:
    post:
      tags: [Routes]
      summary: Mark route as active for current user
      operationId: startRoute
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RouteIdPath'
      responses:
        '200':
          description: Route activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteProgress'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /checkins:
    post:
      tags: [CheckIns]
      summary: Create check-in for a landmark
      description: |
        Business rules:
        - Same user cannot check in to same landmark more than once per day.
        - User must be within 50 meters of the landmark.
      operationId: createCheckIn
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckInRequest'
      responses:
        '201':
          description: Check-in created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInResult'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Duplicate check-in in same day
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate:
                  value:
                    code: DUPLICATE_CHECKIN
                    message: You already checked in at this landmark today.
        '422':
          description: Out of range for check-in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                outOfRange:
                  value:
                    code: OUT_OF_RANGE
                    message: Move closer to this landmark before checking in.
                    details:
                      distanceMeters: 72
                      requiredMeters: 50

    get:
      tags: [CheckIns]
      summary: Get current user's check-in history
      operationId: listMyCheckIns
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: period
          required: false
          schema:
            type: string
            enum: [ALL, LAST_7_DAYS, TODAY]
      responses:
        '200':
          description: Check-in history
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CheckInRecord'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /progress/me:
    get:
      tags: [Progress]
      summary: Get badge and route progress for current user
      operationId: getMyProgress
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProgressResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /admin/landmarks:
    post:
      tags: [Admin]
      summary: Create landmark (admin)
      operationId: createLandmark
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLandmarkRequest'
      responses:
        '201':
          description: Landmark created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Landmark'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /admin/landmarks/{landmarkId}:
    put:
      tags: [Admin]
      summary: Update landmark (admin)
      operationId: updateLandmark
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LandmarkIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLandmarkRequest'
      responses:
        '200':
          description: Landmark updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Landmark'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags: [Admin]
      summary: Delete landmark (admin)
      operationId: deleteLandmark
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LandmarkIdPath'
      responses:
        '204':
          description: Landmark deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /admin/routes:
    post:
      tags: [Admin]
      summary: Create route plan (admin)
      operationId: createRoute
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteRequest'
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePlan'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /admin/routes/{routeId}:
    put:
      tags: [Admin]
      summary: Update route plan (admin)
      operationId: updateRoute
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RouteIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRouteRequest'
      responses:
        '200':
          description: Route updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutePlan'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags: [Admin]
      summary: Delete route plan (admin)
      operationId: deleteRoute
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RouteIdPath'
      responses:
        '204':
          description: Route deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    LandmarkIdPath:
      in: path
      name: landmarkId
      required: true
      schema:
        type: string
    RouteIdPath:
      in: path
      name: routeId
      required: true
      schema:
        type: string

  responses:
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: UNAUTHORIZED
                message: Access token is missing or invalid.

    ForbiddenError:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                code: FORBIDDEN
                message: You do not have permission to access this resource.

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                code: NOT_FOUND
                message: Requested resource does not exist.

  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: student01
        password:
          type: string
          format: password
          example: Passw0rd!

    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string

    AuthTokenResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresInSeconds, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresInSeconds:
          type: integer
          format: int32
          example: 3600
        user:
          $ref: '#/components/schemas/UserSummary'

    UserSummary:
      type: object
      required: [userId, username, displayName, roles]
      properties:
        userId:
          type: string
          example: u01
        username:
          type: string
          example: student01
        displayName:
          type: string
          example: Dong Zhang
        roles:
          type: array
          items:
            type: string
          example: [USER]

    LandmarkCategory:
      type: string
      enum: [historic, nature, food, culture]

    Landmark:
      type: object
      required:
        - id
        - name
        - category
        - address
        - description
        - latitude
        - longitude
        - imageUrl
        - rating
      properties:
        id:
          type: string
          example: l1
        name:
          type: string
          example: Irving House
        category:
          $ref: '#/components/schemas/LandmarkCategory'
        address:
          type: string
        description:
          type: string
        latitude:
          type: number
          format: double
          example: 49.2064
        longitude:
          type: number
          format: double
          example: -122.9094
        imageUrl:
          type: string
          format: uri
        rating:
          type: number
          format: double
          example: 4.7

    RoutePlan:
      type: object
      required: [id, name, distanceKm, durationMinutes, difficulty, landmarkIds]
      properties:
        id:
          type: string
          example: r1
        name:
          type: string
          example: Historic Downtown Walk
        distanceKm:
          type: number
          format: double
          example: 2.5
        durationMinutes:
          type: integer
          format: int32
          example: 45
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        landmarkIds:
          type: array
          items:
            type: string
          example: [l1, l2, l3, l4, l12]

    CreateCheckInRequest:
      type: object
      required: [landmarkId, latitude, longitude]
      properties:
        landmarkId:
          type: string
          example: l1
        latitude:
          type: number
          format: double
          example: 49.2064
        longitude:
          type: number
          format: double
          example: -122.9094
        note:
          type: string
          maxLength: 280
          nullable: true

    CheckInRecord:
      type: object
      required: [id, landmarkId, checkedInAt]
      properties:
        id:
          type: string
          example: c01
        landmarkId:
          type: string
          example: l1
        checkedInAt:
          type: string
          format: date-time
        note:
          type: string
          nullable: true

    RouteProgress:
      type: object
      required: [routeId, routeName, completedStops, totalStops]
      properties:
        routeId:
          type: string
          example: r1
        routeName:
          type: string
          example: Historic Downtown Walk
        completedStops:
          type: integer
          format: int32
          example: 1
        totalStops:
          type: integer
          format: int32
          example: 5
        nextStopLandmarkId:
          type: string
          nullable: true
          example: l2
        nextStopName:
          type: string
          nullable: true
          example: New Westminster Museum
        completed:
          type: boolean
          example: false

    BadgeProgress:
      type: object
      required:
        - visitedCount
        - totalLandmarks
        - bronzeEarned
        - silverEarned
        - goldEarned
        - nextBadgeHint
      properties:
        visitedCount:
          type: integer
          format: int32
          example: 3
        totalLandmarks:
          type: integer
          format: int32
          example: 15
        bronzeEarned:
          type: boolean
        silverEarned:
          type: boolean
        goldEarned:
          type: boolean
        nextBadgeHint:
          type: string
          example: 'Next badge: Bronze (5)'

    CategoryProgress:
      type: object
      required: [category, visitedCount, totalCount]
      properties:
        category:
          $ref: '#/components/schemas/LandmarkCategory'
        visitedCount:
          type: integer
          format: int32
          example: 1
        totalCount:
          type: integer
          format: int32
          example: 4

    UserProgressResponse:
      type: object
      required: [badgeProgress, categoryProgress]
      properties:
        badgeProgress:
          $ref: '#/components/schemas/BadgeProgress'
        categoryProgress:
          type: array
          items:
            $ref: '#/components/schemas/CategoryProgress'
        activeRoute:
          $ref: '#/components/schemas/RouteProgress'
          nullable: true

    CheckInResult:
      type: object
      required: [status, message, checkIn, badgeProgress]
      properties:
        status:
          type: string
          enum: [SUCCESS]
        message:
          type: string
          example: 'Check-in saved. Next badge: Bronze (5)'
        checkIn:
          $ref: '#/components/schemas/CheckInRecord'
        badgeProgress:
          $ref: '#/components/schemas/BadgeProgress'
        routeProgress:
          $ref: '#/components/schemas/RouteProgress'
          nullable: true

    CreateLandmarkRequest:
      type: object
      required:
        - name
        - category
        - address
        - description
        - latitude
        - longitude
        - imageUrl
        - rating
      properties:
        name:
          type: string
        category:
          $ref: '#/components/schemas/LandmarkCategory'
        address:
          type: string
        description:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        imageUrl:
          type: string
          format: uri
        rating:
          type: number
          format: double

    UpdateLandmarkRequest:
      allOf:
        - $ref: '#/components/schemas/CreateLandmarkRequest'

    CreateRouteRequest:
      type: object
      required: [name, distanceKm, durationMinutes, difficulty, landmarkIds]
      properties:
        name:
          type: string
        distanceKm:
          type: number
          format: double
        durationMinutes:
          type: integer
          format: int32
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        landmarkIds:
          type: array
          minItems: 1
          items:
            type: string

    UpdateRouteRequest:
      allOf:
        - $ref: '#/components/schemas/CreateRouteRequest'

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: OUT_OF_RANGE
        message:
          type: string
        details:
          type: object
          additionalProperties: true