# src/main/java/ca/douglas/csis4280/nwtrails/common/ErrorResponse.java
package ca.douglas.csis4280.nwtrails.common;

import java.util.Map;

public record ErrorResponse(String code, String message, Map<String, Object> details) {

    public static ErrorResponse of(String code, String message) {
        return new ErrorResponse(code, message, Map.of());
    }
}